AG-13286 - Add Vue3 package tests.

+++ tests/shared/run.sh
@@ -7,14 +7,24 @@ mode=docker
 interactive=false
 update=false
 production=false
-it_opts=$([[ ${TTY:-} -ne "" ]] && echo "-it" || echo "")
+it_opts=
+passthrough_opts=
 
-passthrough_opts=-n
+function sed_inplace {
+    if [[ $(uname) == "Darwin" ]] ; then
+        sed -i '' "$@"
+    else
+        sed -i'' "$@"
+    fi
+}
 
-while getopts ":eniup" opt; do
+while getopts ":eniupc" opt; do
   case $opt in
     e)
       editor=true
+      ;;
+    c)
+      mode=container
       ;;
     n)
       mode=native
@@ -30,6 +40,7 @@ while getopts ":eniup" opt; do
     i)
       interactive=true
       passthrough_opts="${passthrough_opts} -i"
+      it_opts=-it
       ;;
     \?)
       echo "Invalid option: -$OPTARG" >&2
@@ -44,36 +55,46 @@ done
 shift $((OPTIND - 1))
 
 version=$1
+project=${2:-/project}
 
-if [[ ${mode} == "docker" ]] ; then
+if [[ ${mode} == "container" ]] ; then
+    echo ">>> using prepared temporary project folder..."
+    cd ${project}
+else
+    echo ">>> preparing temporary project folder..."
     repo_dir=$(git rev-parse --show-toplevel)
     project_dir=$(readlink -f $(dirname $0))
     project_script=$(basename $0)
-    temp_dir=$(mktemp -d)
+    project=$(mktemp -d)
 
-    cp -R ${project_dir}/../shared/* $temp_dir/
-    cp -R ${project_dir}/* $temp_dir/
-    cp dist/packages/*.tgz $temp_dir/
+    cp -R ${project_dir}/../shared/* $project/
+    cp -R ${project_dir}/* $project/
+    cp dist/packages/*.tgz $project/
 
     cd ${project_dir}
-    sed -e '/source .*\/run.sh$/r ../shared/run.sh' ${project_dir}/${project_script} >${temp_dir}/run.sh
-    sed -i '' -e '/source .*\/run.sh$/d' ${temp_dir}/run.sh
-    cd ${temp_dir}
+    sed -e '/source .*\/run.sh$/r ../shared/run.sh' ${project_dir}/${project_script} >${project}/run.sh
+    sed_inplace -e '/source .*\/run.sh$/d' ${project}/run.sh
+    cd ${project}
 
-    if ${editor} ; then
-        code . &
-    fi
+    echo ">>> temporary project folder: ${project}"
+fi
+
+if ${editor} ; then
+    code . &
+fi
+
+if [[ ${mode} == "docker" ]] ; then
+    echo ">>> docker run ..."
     port_spec=
     if ${interactive} ; then
-        port_spec=-p 4200:4200
+        port_spec="-p ${dev_port}:${dev_port}"
     fi
-    echo ">>> docker run ..."
     mkdir -p ./npm-cache
     docker run ${it_opts} --rm --ipc=host \
         -v $(pwd):/project \
         $port_spec \
         mcr.microsoft.com/playwright:v1.45.0-jammy \
-        /bin/bash -il /project/run.sh ${passthrough_opts} ${version}
+        /bin/bash -il /project/run.sh -c ${passthrough_opts} ${version} /project
     exitCode=$?
 
     if ${update} ; then
@@ -83,14 +104,12 @@ if [[ ${mode} == "docker" ]] ; then
     exit ${exitCode}
 fi
 
-cd /project
-
 echo ">>> git config"
 git config --global init.defaultBranch latest
 git config --global user.email "me@ag-grid.com"
 git config --global user.name "myself"
 
-npm config set cache /project/.npm-cache
+npm config set cache ${project}/.npm-cache
 install_fw
 
 if ${production} ; then
@@ -104,14 +123,18 @@ git add .
 git commit -m "Initial commit"
 
 for filename in ../patches/* ; do
+    if [ ! -f "$filename" ] ; then
+        continue
+    fi
+
     ext=${filename##*.}
 
     if [[ ${ext} == 'sed' ]] ; then
-        target=$(find src/ -name "$(basename ${filename%.*})")
+        target=$(find src -name "$(basename ${filename%.*})")
         echo ">>> Modifying ${target}"
-        sed -i'' -f $filename $(pwd)/$target
+        sed_inplace -f $filename $(pwd)/$target
     else
-        target=$(find src/ -name "$(basename $filename)")
+        target=$(find src -name "$(basename $filename)")
         echo ">>> Updating ${target}"
         cp $filename $target
     fi
@@ -125,12 +148,12 @@ if ${production} ; then
     export FW_VERSION=production-$FW_VERSION
 fi
 
+build_fw
 if ${interactive} ; then
     serve_fw
     npx playwright test $(${update} && echo "-u" || echo "") || echo "Tests failed"
     /bin/bash -il
 else
-    build_fw
     echo ">>> playwright test"
     npx playwright test $(${update} && echo "-u" || echo "")
 fi
